chore: organise global key inputs in one function
