+++ src/data/mod.rs
@@ -1,3 +1,4 @@
+pub mod fn_utils;
 pub mod io_utils;
 pub mod stores;
 pub mod themes;
